REFACTOR: Consolidate routes for help pages in `help` directory.
